Remove robot.yaml when upgrading package and use pip install instead of git subtree for vendored_deps.
git subtree is basically the same as copying the contents to the repo and having
a command to update it afterwards.

For our use case, using pip install to get the vendored dependencies
directly from the repository seems better as then the code structure
doesn't need to reflect 100% the repository structure (which is
better for our use case).

Also, both require a specific command for updating and git subtree
doesn't seem to bring many benefits.